==> globalprep/spp/v2019/_setup/5a_aggregate_indiv_spp_shps.Rmd <==
---
title: 'Process IUCN spp shapes provided individually'
author: "*Compiled on `r date()` by `r Sys.info()['user']`*"
output: 
  html_document:
    code_folding: hide
    toc: true
    toc_depth: 3
    toc_float: yes
    number_sections: true
    theme: cerulean
    highlight: haddock
    includes: 
      in_header: '../../../../workflow/templates/ohi_hdr.html'
  pdf_document:
    toc: true
---



```{r}
knitr::opts_chunk$set(echo = TRUE, eval=FALSE)
```

``` {r setup, echo = TRUE, message = FALSE, warning = FALSE}

knitr::opts_chunk$set(fig.width = 6, fig.height = 4, fig.path = 'Figs/',
                      echo = TRUE, message = FALSE, warning = FALSE)

library(raster)
library(rgeos)
source('https://raw.githubusercontent.com/oharac/src/master/R/common.R')

library(sf)

source('https://raw.githubusercontent.com/oharac/src/master/R/common.R')  ###
  ### includes library(tidyverse); library(stringr); dir_M points to ohi directory

dir_git <- here()

source(file.path(dir_git, '_setup/common_fxns.R'))


dir_zips <- file.path(dir_M, 'git-annex/globalprep',
                     '_raw_data/iucn_spp/zips_from_iucn_2018')
unzip_dirs <- list.dirs(dir_zips) %>%
  .[. != dir_zips] ### exclude home (dir_zips) from results
  ### These are shapefiles directly from IUCN as individual species map files;
  ### these will be handled differently from those taken from Red List Spatial
  ### Data Download page

dir_shp <- file.path(dir_M, 'git-annex/globalprep',
                     '_raw_data/iucn_spp/d2018-1')

```

# Summary

Gather individually provided species shapefiles into taxa shapefiles for easier processing.

# Data source

IUCN: Gina Ralph direct communication

# Methods

## check new shps vs old shps

``` {r define fxn}
### Define helper function to cleanly get the info from DBF files,
### including colname checks
get_dbf <- function(x) { ### x <- shps_dbfs[1]
  spp_dbf_info <- foreign::read.dbf(x, as.is = TRUE) %>%
    clean_df_names()
  
  ### Individual files call the id number 'iucn_sid'; bli has been fixed to
  ### 'iucn_sid' as well; Red List Data Download files are 'id_no'... So:
  ### fix that here!
  names(spp_dbf_info)[names(spp_dbf_info) == 'id_no'] <- 'iucn_sid'
  
  ### Some files (e.g. bli and individual files) don't have 'subpop':
  ### add it in as NAs.
  if(!'subpop' %in% names(spp_dbf_info)) spp_dbf_info$subpop <- NA
  
  ### some files don't have 'presence' field: if not present or NA,
  ### set to presence = 1.  Some have presence = 0?
  if(!'presence' %in% names(spp_dbf_info)) spp_dbf_info$presence <- 1
  
  spp_dbf_info <- spp_dbf_info %>%
    mutate(presence = ifelse(presence == 0, 1, presence))
  
  return(spp_dbf_info)
}
```

``` {r}

solo_map_files <- list.files(file.path(dir_zips, 'IUCN_moremaps_part1'), 
                        pattern = '.dbf$', full.names = TRUE)
map_by_order_files <- list.files(file.path(dir_zips, 
                                           c('IUCN_maps_byorder_part1',
                                             'Perciformes')),
                           pattern = '.dbf$', full.names = TRUE)

solo_maps <- lapply(solo_map_files, FUN = function(x) {
    foreign::read.dbf(x, as.is = TRUE)
  }) %>%
  setNames(solo_map_files) %>%
  bind_rows(.id = 'dbf') %>%
  select(iucn_sid, sciname = BINOMIAL, dbf)

### read.dbf messes up on reading large integers with trailing zeros.
problem_ids <- c('Acanthurus leucosternon' = 178000)

order_maps <- lapply(map_by_order_files, FUN = function(x) {
    foreign::read.dbf(x, as.is = TRUE)
  }) %>%
  setNames(map_by_order_files) %>%
  bind_rows(.id = 'dbf') %>%
  select(iucn_sid, sciname = BINOMIAL, dbf) %>%
  mutate(iucn_sid = ifelse(is.na(iucn_sid), problem_ids[sciname], iucn_sid))

solo_not_in_orders <- solo_maps %>%
  filter(!iucn_sid %in% order_maps$iucn_sid)

maps_from_iucn <- bind_rows(order_maps, solo_not_in_orders) %>%
  mutate(shp = str_replace(dbf, '.dbf$', '.shp')) %>%
  distinct()

write_csv(maps_from_iucn, file.path(dir_setup, 'int/map_check',
                                    'maps_direct_from_iucn.csv'))
```

## Read spp shapes, trim to +/-180 deg, save as taxa level shapefiles

### Check the validity and bounds of all spp shapefiles

Focus on the non-duplicated shapes from `setup/int/maps_direct_from_iucn.csv`.

``` {r gather all ids and check bounds}

reload <- FALSE

maps_from_iucn <- read_csv(file.path(dir_setup, 'int/map_check',
                                     'maps_direct_from_iucn.csv'))

bounds_check_file <- file.path(dir_setup, 'int/map_check', 
                               'spp_bounds.csv')

if(!file.exists(bounds_check_file) | reload) {
  
  shp_list <- maps_from_iucn$shp %>%
    unique()
  
  ### initialize a data frame 
  all_bounds <- vector('list', length = length(shp_list)) %>%
    setNames(shp_list)
  
  for(shp in shp_list) { ### shp <- shp_list[1]
    shp_sf <- read_sf(shp)
    spp_list <- shp_sf$iucn_sid
    cat_msg('Processing ', basename(shp), ' with ', 
            length(unique(shp_sf$iucn_sid)), ' spp')

    system.time({
      spp_bounds_list <- parallel::mclapply(spp_list, mc.cores = 12, 
        FUN = function(spp) { ### spp <- spp_list[1]
          spp_sf <- shp_sf %>%
            filter(iucn_sid == spp)
        
          spp_bbox <- st_bbox(spp_sf)
          spp_valid <- st_is_valid(spp_sf)
        
          spp_df <- data.frame(iucn_sid = spp_sf$iucn_sid,
                               spp_shp = shp,
                               valid = spp_valid,
                               xmin = spp_bbox$xmin,
                               xmax = spp_bbox$xmax,
                               ymin = spp_bbox$ymin,
                               ymax = spp_bbox$ymax) %>%
            mutate(exceeds_bounds = (xmin < -180 | xmax > 180))
        
      }) ### end of mclapply
    }) ### end of system.time
    all_bounds[[shp]] <- bind_rows(spp_bounds_list)
  } ### end of for-loop across shapes

  spp_bounds_df <- bind_rows(all_bounds) %>%
    distinct()
  
  write_csv(spp_bounds_df, bounds_check_file)
  
} else {
  spp_bounds_df <- read_csv(bounds_check_file)
}

DT::datatable(spp_bounds_df %>% mutate(spp_shp = basename(spp_shp)))
```

### Attempt to trim problem files

After trimming, overwrite the original shapefiles (we still have the ZIPs for *original* originals)

``` {r attempt to trim indiv shapefiles and overwrite}
### This will try to fix each original shapefile from IUCN by clipping
### bounds to +/- 180 (only for those that exceed those bounds).

spp_bounds_df <- read_csv(bounds_check_file)
cat_msg('Spp shapes that exceed x bounds: ', sum(spp_bounds_df$exceeds_bounds))

shps_to_fix <- spp_bounds_df %>%
  filter(exceeds_bounds)

skip <- FALSE

if(nrow(shps_to_fix) > 0 & skip == FALSE) {
  shps_list <- shps_to_fix$spp_shp %>% 
    unique() %>% 
    sort(decreasing = TRUE)
  
  spp_bounds_list <- parallel::mclapply(seq_along(shps_list),
                                          mc.cores = 12, FUN = function(i) {
    ### i <- 1
    shp_file <- shps_list[i]
    cat_msg(i, '. processing ', basename(shp_file))
    spp_sf <- read_sf(shp_file)
    spp_valid <- shps_to_fix %>%
      filter(spp_shp == shp_file) %>%
      .$valid %>%
      all()
    
    if(!spp_valid) {
      cat_msg(shp_file, ': invalid geometry; buffering zero distance')
      spp_sf <- st_buffer(spp_sf, 0)
    }
    
    ### if sf version (st_crop) fails, try with raster version (raster::crop)
    tryCatch(spp_sf_trim <- clip_to_globe(spp_sf), 
             finally = spp_sf_trim <- clip_to_globe_sp(spp_sf))
  
    write_sf(spp_sf_trim, shp_file, driver = 'ESRI Shapefile')
  
  }) ### end of mclapply
} else {
  cat_msg(nrow(shps_to_fix), ' shapes that exceed bounds; skipping for now')
}

```

## Combine spp shapes into taxa shapes

Here we combine the standalone species maps with the family and order level mapsets; there are a few standalones that aren't included in the others for some reason.

### Identify the species and taxa from the dbfs

``` {r generate map list from dbfs}

maps_from_iucn <- read_csv(file.path(dir_setup, 'int/map_check',
                                     'maps_direct_from_iucn.csv'))

### Notes on presence, origin, seasonal fields:
### * presence = 5 is extinct; 4 = probably extinct; others are extant-ish or 
###   uncertain. We will drop field and include all polygons for now.
### * origin is native, introduced, etc.  We will drop this field and not
###   worry about origin.
### * seasonal is breeding/non-breeding/passage.  We will drop this field
###   and not worry about seasonality.

### filter to marine habitat species (to drop terrestrial reptiles e.g.)
marine_spp_depths <- read_csv(file.path(dir_git, '_data',
                                     sprintf('spp_marine_from_api_%s.csv', 
                                             api_version)),
                           col_types = 'dcc') %>%
  select(iucn_sid, max_depth)

### Find taxonomic info and attach
spp_taxa <- read_csv(file.path(dir_o_anx, 'iucn',
                               sprintf('spp_info_from_api_%s.csv', api_version)),
                     col_types = 'dccccc____') %>%
  mutate(phylum = tolower(phylum),
         class  = tolower(class),
         order  = tolower(order),
         family = tolower(family))

marine_map_info <- maps_from_iucn %>%
  inner_join(marine_spp_depths, by = 'iucn_sid') %>%
  inner_join(spp_taxa, by = 'iucn_sid') %>%
  mutate(taxon = case_when(order  == 'perciformes'    ~ family,
                           class  == 'actinopterygii' ~ order,
                           phylum == 'chordata'       ~ class,
                           TRUE                       ~ phylum))

write_csv(marine_map_info, 
          file.path(dir_setup, 'int/map_check', 
                    'marine_spp_processed.csv'))
```

### Gather the species by taxa and save as a taxa-level shapefile

Note this stops out if the bounds list is not resolved.  Any problem species will need to be examined manually and fixed, then the above chunks (currently set as eval = FALSE) will need to be rerun to update the spp_bounds.csv file.

Other things fixed here:

* changes id_no field to iucn_sid where applicable
* changes binomial to sciname where applicable
* adds presence field where applicable, and sets any presence == 0 to 1
* adds subpop field where applicable, and sets to NA if none originally present
* for freakin' Acanthurus leucosternon, manually sets species ID to 178000

``` {r gather ids and read into a single shapefile}

reload <- FALSE

spp_probs <- read_csv(bounds_check_file) %>%
  filter(exceeds_bounds)
if(nrow(spp_probs) > 0) stop('Still species maps that exceed +/- 180 degrees!')

maps_shps <- read_csv(file.path(dir_setup, 'int/map_check',
                                'marine_spp_processed.csv'))
  
taxa <- maps_shps$taxon %>% unique() %>% sort()
  
for(taxon_gp in taxa) { ### taxon_gp <- taxa[1]
  taxon_spp <- maps_shps %>%
    filter(taxon == taxon_gp)
  taxon_shps <- taxon_spp$shp %>% unique()
  
  cat_msg('Processing ', taxon_gp, ' with ', nrow(taxon_spp), 
          ' spp in ', length(taxon_shps), ' shapefile(s)')
  
  outfile <- file.path(dir_shp, paste0(taxon_gp, '.shp'))
  
  if(!file.exists(outfile) | reload == TRUE) {
  
    shps_list <- lapply(taxon_shps, read_sf)
    
    polys_all <- do.call(rbind, shps_list) %>%
      ### allows to pass a *list* of arguments to a function separately!
      clean_df_names()
    if(!'sciname' %in% names(polys_all) & 'binomial' %in% names(polys_all)) {
      polys_all <- polys_all %>%
        rename(sciname = binomial)
    }
    if('id_no' %in% names(polys_all)) {
      polys_all <- polys_all %>%
        rename(iucn_sid = id_no)
    }
    polys_all <- polys_all %>%
      mutate(iucn_sid = ifelse(sciname == 'Acanthurus leucosternon', 
                               178000, iucn_sid),
             iucn_sid = as.numeric(iucn_sid)) %>%
      filter(iucn_sid %in% taxon_spp$iucn_sid)
    
    ### clean up the attributes of resulting sf object
    if(!'subpop' %in% names(polys_all)) {
      polys_all$subpop <- NA_character_
      ### if shape doesn't have subpop column, add it as NA
    }
    if(!'presence' %in% names(polys_all)) {
      polys_all <- polys_all %>%
        mutate(presence = 1)
    }
    
    polys_all <- polys_all %>%
      mutate(presence = ifelse(presence == 0, 1, presence))
    
    if(any(polys_all$iucn_sid > 1e8)) {
      cat_msg("don't freak out about the error message - the field seems fine")
    }
    st_write(polys_all, outfile, driver = 'ESRI Shapefile',
             delete_layer = TRUE)
    
  } else {
    cat_msg('Shapefile found: ', outfile)
  }
}

```
